backend/requirements.txt
Django==5.2.13
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
python-dotenv==1.0.1
django-cors-headers==4.7.0
gunicorn==23.0.0
mysqlclient==2.2.4
